refactor: Rename `lib/plugins/aws/info/getStackInfo.js`
