+++ lib/plugins/aws/info/get-stack-info.js
