+++ lib/plugins/aws/info/index.js
@@ -4,7 +4,7 @@ const BbPromise = require('bluebird');
 const { progress, style } = require('@serverless/utils/log');
 const writeServiceOutputs = require('../../../cli/write-service-outputs');
 const validate = require('../lib/validate');
-const getStackInfo = require('./getStackInfo');
+const getStackInfo = require('./get-stack-info');
 const getResourceCount = require('./get-resource-count');
 const getApiKeyValues = require('./get-api-key-values');
 const display = require('./display');
