+++ test/unit/lib/plugins/aws/info/get-stack-info.test.js
